requirements.txt
streamlit
feedparser
openai
python-dotenv
requests
